+++ spec/entities/road_spec.rb
@@ -31,4 +31,16 @@ describe Road do
   it 'is not adjacent to other spots' do
     expect(road).not_to be_adjacent_to(other_spot)
   end
+
+  it "can't be created if spots out of map" do
+    expect do
+      Road.new(from: 1, to: 100, owner: player)
+    end.to raise_error(ArgumentError, 'Invalid from or to param')
+  end
+
+  it "can't be created if spots don't border" do
+    expect do
+      Road.new(from: 1, to: 3, owner: player)
+    end.to raise_error(ArgumentError, "Spots don't border")
+  end
 end